♻️ Split updater backup path resolution.
